+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This file contains "important" commits, small ones will probably not be added.
 
 ### 200227 - 1 [2.0.25]
+- New: languages switching right from the command prompt
 - Fixed: there were some issues when switching between languages
 - New: windows opening animation
 - Improve: design of side menu labels, and command prompt input